fix : post servisi ile postProcess,postMedia ve comment servisleri birleştirildi fix : gereksiz dosyalar temizlendi fix : Boolean değişkenler primitive dönüşümü
