+++ src/main/java/org/gelecekbilimde/scienceplatform/auth/model/Role.java
@@ -31,7 +31,7 @@ public class Role {
 	private String description;
 
 	@Column(columnDefinition = "boolean default false")
-	private Boolean isDefault;
+	private boolean isDefault = false;
 
 	@ManyToMany(fetch = FetchType.EAGER)
 	@JoinTable(name = "role_permission", joinColumns = @JoinColumn(name = "role_id"), inverseJoinColumns = @JoinColumn(name = "permission_id"))
