--- src/main/java/org/gelecekbilimde/scienceplatform/comment/dto/domain/domain/CommentDomain.java
@@ -1,4 +1,0 @@
-package org.gelecekbilimde.scienceplatform.comment.dto.domain.domain;
-
-public class CommentDomain {
-}
