--- src/main/java/org/gelecekbilimde/scienceplatform/comment/dto/domain/request/CommentRequest.java
@@ -1,4 +1,0 @@
-package org.gelecekbilimde.scienceplatform.comment.dto.domain.request;
-
-public class CommentRequest {
-}
