--- src/main/java/org/gelecekbilimde/scienceplatform/comment/dto/domain/response/CommentResponse.java
@@ -1,4 +1,0 @@
-package org.gelecekbilimde.scienceplatform.comment.dto.domain.response;
-
-public class CommentResponse {
-}
