--- src/main/java/org/gelecekbilimde/scienceplatform/comment/model/Comment.java
@@ -1,31 +1,0 @@
-package org.gelecekbilimde.scienceplatform.comment.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.gelecekbilimde.scienceplatform.post.model.Post;
-
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-
-@Table(name = "comment")
-public class Comment {
-	@Id
-	@GeneratedValue
-	private Long id;
-
-	@Column(columnDefinition = "text", nullable = false)
-	private String comment;
-
-	@Column(columnDefinition = "integer", nullable = false)
-	private Integer likeCount;
-
-	// todo : oneToMany olmalı
-	@ManyToMany(mappedBy = "comments")
-	private Set<Post> posts = new HashSet<>();
-
-}
