+++ src/main/java/org/gelecekbilimde/scienceplatform/common/Response.java
@@ -12,10 +12,6 @@ import java.util.*;
 @Builder
 public class Response<T> {
 
-	private final HttpStatus statusText;
-	private final Integer status;
-
-
 	@JsonInclude(JsonInclude.Include.NON_NULL)
 	private T list;
 
@@ -26,14 +22,10 @@ public class Response<T> {
 	private String responseCode;
 
 	public static final Response<Void> NO_CONTENT = Response.<Void>builder()
-		.statusText(HttpStatus.NO_CONTENT)
-		.status(HttpStatus.NO_CONTENT.value())
 		.build();
 
 	public static <T> Response<T> ok(final T response) {
 		return Response.<T>builder()
-			.statusText(HttpStatus.OK)
-			.status(HttpStatus.OK.value())
 			.list(response)
 			.responseCode(UUID.randomUUID().toString())
 			.build();
@@ -42,8 +34,6 @@ public class Response<T> {
 
 	public static <T> Response<T> create(final T response) {
 		return Response.<T>builder()
-			.statusText(HttpStatus.CREATED)
-			.status(HttpStatus.CREATED.value())
 			.list(response)
 			.responseCode(UUID.randomUUID().toString())
 			.build();
