+++ src/main/java/org/gelecekbilimde/scienceplatform/config/JwtAuthenticationFilter.java
@@ -43,7 +43,6 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 		final String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
 		final String jwt;
 		final String userName;
-		final Integer userId;
 		final String roleName;
 
 		if (null == authHeader || !authHeader.startsWith("Bearer ")) {
@@ -54,7 +53,6 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 		jwt = authHeader.substring(7);
 		userName 	= jwtService.extractSubject(jwt);
 		roleName  	= jwtService.extractClaim(jwt,"role").toString();
-		userId 		= Integer.parseInt(jwtService.extractClaim(jwt,"userId").toString());
 
 		if (null != SecurityContextHolder.getContext().getAuthentication()){
 			filterChain.doFilter(request,response);
