+++ src/main/java/org/gelecekbilimde/scienceplatform/post/dto/domain/AdminPostListDomain.java
@@ -3,7 +3,7 @@ package org.gelecekbilimde.scienceplatform.post.dto.domain;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.SuperBuilder;
-import org.gelecekbilimde.scienceplatform.postprocess.enums.PostProcessEnum;
+import org.gelecekbilimde.scienceplatform.post.enums.PostProcessEnum;
 
 import java.util.List;
 
@@ -20,9 +20,9 @@ public class AdminPostListDomain {
 	private PostProcessEnum lastProcess;
 	private List<String> label;
 	private Integer likeCount ;
-	private Boolean active;
-	private Boolean copyrightControl;
-	private Boolean typoControl;
-	private Boolean dangerousControl;
+	private boolean active;
+	private boolean copyrightControl;
+	private boolean typoControl;
+	private boolean dangerousControl;
 
 }
