+++ src/main/java/org/gelecekbilimde/scienceplatform/post/dto/domain/PostDomain.java
@@ -3,7 +3,7 @@ package org.gelecekbilimde.scienceplatform.post.dto.domain;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.SuperBuilder;
-import org.gelecekbilimde.scienceplatform.postprocess.enums.PostProcessEnum;
+import org.gelecekbilimde.scienceplatform.post.enums.PostProcessEnum;
 
 import java.util.List;
 
@@ -21,10 +21,10 @@ public class PostDomain {
 	private PostProcessEnum lastProcess;
 	private List<String> label;
 	private Integer likeCount ;
-	private Boolean active;
-	private Boolean copyrightControl;
-	private Boolean typoControl;
-	private Boolean dangerousControl;
+	private boolean active;
+	private boolean copyrightControl;
+	private boolean typoControl;
+	private boolean dangerousControl;
 
 	private Long userId;
