+++ src/main/java/org/gelecekbilimde/scienceplatform/post/dto/domain/PostMediaDomain.java
@@ -1,0 +1,5 @@
+package org.gelecekbilimde.scienceplatform.post.dto.domain;
+
+public class PostMediaDomain {
+
+}
