+++ src/main/java/org/gelecekbilimde/scienceplatform/post/dto/response/AdminPostListResponse.java
@@ -3,10 +3,8 @@ package org.gelecekbilimde.scienceplatform.post.dto.response;
 import lombok.Builder;
 import lombok.Getter;
 
-import org.gelecekbilimde.scienceplatform.comment.dto.domain.response.CommentResponse;
-import org.gelecekbilimde.scienceplatform.postmedia.dto.response.PostMediaResponse;
 import org.gelecekbilimde.scienceplatform.user.dto.response.UserResponse;
-import org.gelecekbilimde.scienceplatform.postprocess.enums.PostProcessEnum;
+import org.gelecekbilimde.scienceplatform.post.enums.PostProcessEnum;
 
 import java.util.List;
 
@@ -22,11 +20,10 @@ public class AdminPostListResponse {
 	private PostProcessEnum lastProcess;
 	private List<String> label;
 	private Integer likeCount;
-	private Boolean active;
-	private Boolean copyrightControl;
-	private Boolean typoControl;
-	private Boolean dangerousControl;
+	private boolean active;
+	private boolean copyrightControl;
+	private boolean typoControl;
+	private boolean dangerousControl;
 	private UserResponse user;
-	private CommentResponse comment;
 	private PostMediaResponse media;
 }
