+++ src/main/java/org/gelecekbilimde/scienceplatform/post/dto/response/PostMediaResponse.java
@@ -1,0 +1,5 @@
+package org.gelecekbilimde.scienceplatform.post.dto.response;
+
+public class PostMediaResponse {
+
+}
