+++ src/main/java/org/gelecekbilimde/scienceplatform/post/dto/response/PostResponse.java
@@ -3,9 +3,7 @@ package org.gelecekbilimde.scienceplatform.post.dto.response;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.SuperBuilder;
-import org.gelecekbilimde.scienceplatform.comment.dto.domain.response.CommentResponse;
-import org.gelecekbilimde.scienceplatform.postmedia.dto.response.PostMediaResponse;
-import org.gelecekbilimde.scienceplatform.postprocess.enums.PostProcessEnum;
+import org.gelecekbilimde.scienceplatform.post.enums.PostProcessEnum;
 import org.gelecekbilimde.scienceplatform.user.dto.response.UserResponse;
 
 import java.util.List;
@@ -23,13 +21,12 @@ public class PostResponse {
 	private PostProcessEnum lastProcess;
 	private List<String> label;
 	private Integer likeCount ;
-	private Boolean active;
-	private Boolean copyrightControl;
-	private Boolean typoControl;
-	private Boolean dangerousControl;
+	private boolean active;
+	private boolean copyrightControl;
+	private boolean typoControl;
+	private boolean dangerousControl;
 
 	private UserResponse user;
-	private CommentResponse comment;
 	private PostMediaResponse media;
 
 }
