+++ src/main/java/org/gelecekbilimde/scienceplatform/post/enums/PostProcessEnum.java
@@ -1,0 +1,14 @@
+package org.gelecekbilimde.scienceplatform.post.enums;
+
+public enum PostProcessEnum {
+
+	CREATE,
+	CONTROL_TYPO,
+	CONTROL_TRUTH,
+	CONTROL_COPYRIGHT,
+
+	REQUEST_EDIT,
+	REJECT,
+
+	PUBLISH,
+}
