+++ src/main/java/org/gelecekbilimde/scienceplatform/post/repository/PostProcessRepository.java
@@ -1,0 +1,10 @@
+package org.gelecekbilimde.scienceplatform.post.repository;
+
+import org.gelecekbilimde.scienceplatform.post.model.PostProcess;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PostProcessRepository extends JpaRepository<PostProcess, Integer> {
+	Optional <PostProcess> findByPostProcess(Integer id);
+}
