--- src/main/java/org/gelecekbilimde/scienceplatform/postmedia/dto/domain/PostMediaDomain.java
@@ -1,5 +1,0 @@
-package org.gelecekbilimde.scienceplatform.postmedia.dto.domain;
-
-public class PostMediaDomain {
-
-}
