--- src/main/java/org/gelecekbilimde/scienceplatform/postmedia/dto/response/PostMediaResponse.java
@@ -1,5 +1,0 @@
-package org.gelecekbilimde.scienceplatform.postmedia.dto.response;
-
-public class PostMediaResponse {
-
-}
