--- src/main/java/org/gelecekbilimde/scienceplatform/postprocess/dto/domain/PostProcessDomain.java
@@ -1,22 +1,0 @@
-package org.gelecekbilimde.scienceplatform.postprocess.dto.domain;
-
-import lombok.Data;
-import org.gelecekbilimde.scienceplatform.postprocess.enums.PostProcessEnum;
-
-import java.time.LocalDateTime;
-
-
-@Data
-
-public class PostProcessDomain {
-
-
-	private String header;
-	private String slug;
-	private String content;
-
-	private PostProcessEnum lastProcess;
-	private String message;
-	private LocalDateTime createdDate;
-
-}
