--- src/main/java/org/gelecekbilimde/scienceplatform/postprocess/enums/PostProcessEnum.java
@@ -1,14 +1,0 @@
-package org.gelecekbilimde.scienceplatform.postprocess.enums;
-
-public enum PostProcessEnum {
-
-	CREATE,
-	CONTROL_TYPO,
-	CONTROL_TRUTH,
-	CONTROL_COPYRIGHT,
-
-	REQUEST_EDIT,
-	REJECT,
-
-	PUBLISH,
-}
